backend/requirements/base.txt
alembic==1.12.0
bcrypt==4.1.2
cryptography==41.0.4
fastapi==0.103.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.7
pydantic==2.3.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
slowapi==0.1.9
sqlalchemy==2.0.20
structlog==23.1.0
uvicorn[standard]==0.23.0
structlog==23.1.0
pydicom==2.4.3
pillow==10.0.1
boto3==1.28.57
opencv-python-headless==4.8.1.78
numpy==1.24.3
celery[redis]==5.3.4
flower==2.0.1
redis==4.6.0